requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
psycopg2-binary
Pillow
djangorestframework-simplejwt
